# docs/api-definition.yml
openapi: 3.0.0
info:
  version: '1.0.0'
  title: 'Food-Delivery API'
  description: |
    Food-Delivery API based on the OpenAPI 3.0 specification.

    Useful links:
    - [Food-Delivery API Repository](https://github.com/evanigwilo/food-delivery-api)
    - [The API Definition](https://github.com/evanigwilo/food-delivery-api/tree/main/docs/api-definition.yml)
...
paths:
...
  /order/create:
    post:
      tags:
        - Order
      summary: Creates an order
      description: This can only be done by an authenticated user.
      operationId: createOrder
      requestBody:
        required: true
        description: Order description
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Order to create was successful
          content:
            application/json:
              example:
                code: 'SUCCESS'
                message: 'Order created successfully.'
                order:
                  id: '032142ab-9f0d-443e-9347-d24ec084e985'
                  status: 'PENDING'
                  amount: 70.00
                  total: 2
                  items:
                    - id: 'f250c14c-289d-4159-bd21-35828668bf7e'
                      food_name: 'Small Fresh Tuna'
                      food_price: 20.00
                      count: 1
                      menu: 'Breakfast Menu'
                      restaurant_name: 'Tikkaway'
                      restaurant_address: '49010 Mackenzie Isle'
                      restaurant_country: 'United States'
                    - id: '32739a7f-4667-402e-9d44-3e6493a4a564'
                      food_name: 'Pizza'
                      food_price: 50.00
                      count: 1
                      menu: 'Dinner Menu'
                      restaurant_name: 'Tikkaway'
                      restaurant_address: '49010 Mackenzie Isle'
                      restaurant_country: 'United States'

        '400':
          description: ''
          content:
            application/json:
              examples: # Multiple examples
                Invalid Order:
                  description: 'Order properties is not supplied or is invalid'
                  value:
                    code: 'VALIDATION_ERROR'
                    message: 'Invalid Order.'
                Invalid Item(s) Or Count:
                  value:
                    code: 'FAILED'
                    message: 'Order not created - Invalid item(s) or count'

        '401':
          $ref: '#/components/responses/Unauthorized'
  /order/{orderId}:
    get:
      tags:
        - Order
      summary: Get an order by id
      operationId: getOrder
      parameters:
        - $ref: '#/components/parameters/Order'
      responses:
        '200':
          description: Returns order with the specified id
          content:
            application/json:
              schema:
                example:
                  code: 'SUCCESS'
                  message: 'Order found.'
                  order:
                    id: '032142ab-9f0d-443e-9347-d24ec084e985'
                    status: 'PENDING'
                    amount: 70.00
                    total: 2
                    items:
                      - id: 'f250c14c-289d-4159-bd21-35828668bf7e'
                        food_name: 'Small Fresh Tuna'
                        food_price: 20.00
                        count: 1
                        menu: 'Breakfast Menu'
                        restaurant_name: 'Tikkaway'
                        restaurant_address: '49010 Mackenzie Isle'
                        restaurant_country: 'United States'
                      - id: '32739a7f-4667-402e-9d44-3e6493a4a564'
                        food_name: 'Pizza'
                        food_price: 50.00
                        count: 1
                        menu: 'Dinner Menu'
                        restaurant_name: 'Tikkaway'
                        restaurant_address: '49010 Mackenzie Isle'
                        restaurant_country: 'United States'
        '204':
          description: |
            Order not found.

            Returns no content.
    patch:
      tags:
        - Order
      summary: Pay for an order by id
      description: This can only be done by an authenticated user who created the order.
      operationId: updateOrder
      parameters:
        - $ref: '#/components/parameters/Order'
      responses:
        '200':
          description: Order was paid successfully
          content:
            application/json:
              example:
                code: 'SUCCESS'
                message: 'Order Paid.'
                balance: 200
                payment: 50

        '400':
          description: Order payment was not unsuccessful
          content:
            application/json:
              examples:
                Order Already Paid:
                  value:
                    code: 'FAILED'
                    message: 'Order already paid.'
                    payment:
                      id: '032142ab-9f0d-443e-9347-d24ec084e985'
                      status: 'PAID'
                      amount: 70.00
                      total: 2
                Order Not Found:
                  value:
                    code: 'FAILED'
                    message: 'Order not found - Invalid order identifier or user'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  schemas:
...
    Order:
      type: object
      required:
        - orders
      properties:
        orders:
          type: array
          items:
            type: string
          example:
            - foodId: 'f250c14c-289d-4159-bd21-35828668bf7e'
              count: 1
            - foodId: '32739a7f-4667-402e-9d44-3e6493a4a564'
              count: 1
...
  responses:
    Unauthorized:
      description: 'User is not authorized'
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: string
                example: UNAUTHENTICATED
              message:
                type: string
                example: User not authenticated
  parameters:
...
    Order:
      name: orderId
      in: path
      description: Order matching the specified id
      required: true
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid # cookie name

# Apply cookie auth globally to all operations
security:
  - cookieAuth: []